+++ Witness.java
@@ -2,8 +2,8 @@
 public class Witness extends Person {
     protected String witnessStatement;
 
-    public Witness(String witnessState, String fName, String lname, int Age, int Height, int Weight, String Gender,
-            String Address, String Phone, String DateOfBirth, boolean IsAdult, int id) {
+    public Witness(String witnessState, String fName, String lname, long Age, long Height, long Weight, String Gender,
+            String Address, String Phone, String DateOfBirth, boolean IsAdult, long id) {
         super(fName, lname, Age, Height, Weight, Gender, Address, Phone, DateOfBirth, IsAdult, id);
         this.witnessStatement = witnessState;
